Split ShareSidebar/Sidebar into multiple files
Sometimes 800 lines is a mess, especially with all them hooks
running around.
