+++ app/src/userpages/components/ShareSidebar/Sidebar.jsx
@@ -1,479 +1,34 @@
-import React, { useCallback, useState, useRef, useEffect, useMemo, useContext, useReducer } from 'react'
+import React, { useCallback, useState, useRef, useEffect, useContext } from 'react'
 import { connect } from 'react-redux'
 import { Translate, I18n } from 'react-redux-i18n'
 import cx from 'classnames'
-import startCase from 'lodash/startCase'
-import { useSpring, useTransition, animated } from 'react-spring'
+import { useTransition, animated } from 'react-spring'
 
 import useIsMounted from '$shared/hooks/useIsMounted'
 import SelectInput from '$ui/Select'
-import Errors from '$ui/Errors'
-import RadioButtonGroup from './RadioButtonGroup'
 import Button from '$shared/components/Button'
 import { SidebarContext } from '$shared/components/Sidebar/SidebarProvider'
-import Checkbox from '$shared/components/Checkbox'
-import SvgIcon from '$shared/components/SvgIcon'
 import useUniqueId from '$shared/hooks/useUniqueId'
-import TextInput from '$ui/Text'
-import { isFormElement } from '$shared/utils/isEditableElement'
-import {
-    getResourcePermissions,
-    addResourcePermission,
-    removeResourcePermission,
-} from '$userpages/modules/permission/services'
 
 import * as State from './state'
 import styles from './ShareSidebar.pcss'
-import useMeasure from './useMeasure'
 import CopyLink from './CopyLink'
-import Tooltip from '$shared/components/Tooltip'
+import InputNewShare from './InputNewShare'
+import UserPermissions from './UserPermissions'
+import useAsyncCallbackWithState from './hooks/useAsyncCallbackWithState'
+import usePrevious from './hooks/usePrevious'
+import useSlideIn from './hooks/useSlideIn'
+import useUserPermissionState from './hooks/useUserPermissionState'
+import usePermissionsLoader from './hooks/usePermissionsLoader'
+import savePermissions from './utils/savePermissions'
 
 const options = ['onlyInvited', 'withLink']
-
-function useAsyncCallbackWithState(callback) {
-    const [state, setState] = useState({
-        hasStarted: false,
-        isLoading: false,
-        error: undefined,
-        result: undefined,
-    })
-
-    const isMounted = useIsMounted()
-    const { isLoading } = state
-
-    const run = useCallback(async () => {
-        let error
-        let result
-        if (!isMounted()) { return }
-        if (isLoading) { return } // already loading
-
-        setState((s) => ({
-            ...s,
-            isLoading: true,
-            hasStarted: true,
-        }))
-
-        try {
-            result = await callback()
-        } catch (err) {
-            error = err
-        } finally {
-            // only do something if mounted
-            if (isMounted()) {
-                setState({
-                    error,
-                    result,
-                    isLoading: false,
-                    hasStarted: true,
-                })
-            }
-        }
-    }, [isLoading, isMounted, callback])
-    return [state, run]
-}
-
-function usePrevious(value) {
-    const ref = useRef()
-    useEffect(() => {
-        ref.current = value
-    }, [value])
-    return ref.current
-}
-
-/**
- * Transitions element from height: 0 to measured height.
- */
-
-function useSlideIn({ isVisible } = {}) {
-    isVisible = !!isVisible
-    const previousIsVisible = !!usePrevious(isVisible)
-
-    const [bind, { height }] = useMeasure() // attach bind to element to measure
-    const justChanged = previousIsVisible !== isVisible
-
-    const targetHeight = isVisible ? height : 0
-    const selectedHeight = targetHeight
-
-    const [, forceUpdate] = useReducer((x) => x + 1, 0)
-    useEffect(() => {
-        if (justChanged) {
-            // needs second render to perform transition correctly
-            forceUpdate()
-        }
-    }, [justChanged, forceUpdate])
-
-    const style = useSpring({
-        height: selectedHeight,
-        config: {
-            mass: 1,
-            friction: 62,
-            tension: 700,
-            precision: 0.00001,
-        },
-    })
-
-    return [bind, style]
-}
-
-/**
- * Input for adding new users.
- */
-
-function InputNewShare({ currentUser, onChange, canShareToUser }) {
-    const [value, setValue] = useState('')
-    const onChangeValue = useCallback((e) => {
-        setValue(e.target.value.trim())
-    }, [setValue])
-    const onAdd = useCallback(() => {
-        onChange(value)
-        setValue('')
-    }, [value, onChange])
-
-    const uid = useUniqueId('InputNewShare')
-
-    function getShareToUserError({ currentUser, userId }) {
-        if (!State.isValidUserId(userId)) { return I18n.t('share.error.invalidUserError') }
-        if (userId === 'anonymous') { return I18n.t('share.error.anonymousUserError') }
-        if (userId === currentUser) { return I18n.t('share.error.currentUserError') }
-    }
-
-    const error = getShareToUserError({
-        currentUser,
-        userId: value,
-    })
-
-    const isValid = canShareToUser(value)
-    const [trySubmit, setTrySubmit] = useState(false)
-
-    // only show validation when not focussed
-    const [shouldShowValidation, setShouldShowValidation] = useState(false)
-    const onBlur = useCallback(() => {
-        setShouldShowValidation(true)
-    }, [])
-
-    const onFocus = useCallback(() => {
-        setTrySubmit(false)
-        setShouldShowValidation(false)
-    }, [])
-
-    const onKeyDown = useCallback((event) => {
-        // try add user on enter
-        if (event.key === 'Enter') {
-            setTrySubmit(true)
-        } else {
-            setTrySubmit(false)
-            setShouldShowValidation(false)
-        }
-    }, [])
-
-    // only add user on enter if valid
-    const shouldTrySubmit = !!(value && isValid && trySubmit)
-    const onAddRef = useRef()
-    onAddRef.current = onAdd
-    // trigger onAdd in effect so value/validity state has chance to update
-    useEffect(() => {
-        if (!shouldTrySubmit) { return }
-        onAddRef.current()
-        setTrySubmit(false)
-    }, [shouldTrySubmit, onAddRef])
-
-    const showValidationError = shouldShowValidation && value && !isValid
-
-    return (
-        <div className={styles.InputNewShare}>
-            <label htmlFor={uid}>{I18n.t('auth.labels.address')}</label>
-            <TextInput
-                id={uid}
-                className={styles.input}
-                placeholder={I18n.t('modal.shareResource.enterEmailAddress')}
-                value={value}
-                onChange={onChangeValue}
-                onFocus={onFocus}
-                onBlur={onBlur}
-                autoComplete="email"
-                invalid={showValidationError}
-                onKeyDown={onKeyDown}
-            />
-            <Button
-                kind="secondary"
-                onClick={onAdd}
-                disabled={!isValid}
-                className={styles.button}
-            >
-                <SvgIcon name="plus" className={styles.plusIcon} />
-            </Button>
-            {showValidationError && <Errors>{error}</Errors>}
-        </div>
-    )
-}
-
-/**
- * Individual User's Permissions UI
- */
-
-function UserPermissions({
-    resourceType,
-    userId,
-    userPermissions,
-    updatePermission,
-    removeUser,
-    className,
-    onSelect,
-    isSelected,
-    isCurrentUser,
-    error,
-}) {
-    const selectedGroupName = State.findPermissionGroupName(resourceType, userPermissions)
-    // custom handling:
-    // if user edits permissions after clicking a preset, preset will be set to custom (if config doesn't match another preset)
-    // when user actively clicks the custom tab, it will use whatever permissions were currently set
-
-    const isCustom = State.isCustom(resourceType, selectedGroupName, userPermissions)
-
-    const [bind, permissionControlsStyle] = useSlideIn({ isVisible: isSelected })
-
-    const permissionGroupOptions = Object.keys(State.getPermissionGroups(resourceType)).filter((name) => name !== 'default')
-
-    const onClick = useCallback((event) => {
-        if (isFormElement(event.target)) { return }
-        // toggle open state
-        if (isSelected) {
-            return onSelect()
-        }
-        onSelect(userId)
-    }, [isSelected, onSelect, userId])
-
-    // eslint-disable-next-line max-len
-    /* eslint-disable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events, jsx-a11y/no-noninteractive-element-interactions */
-    return (
-        <div
-            className={cx(styles.userPermissions, className, {
-                [styles.isSelected]: isSelected,
-            })}
-        >
-            <div className={styles.permissionsHeader} onClick={onClick}>
-                <div className={styles.permissionsHeaderTitle}>
-                    <h4 title={userId}>
-                        {userId}
-                        {!!isCurrentUser && (' (You)')}
-                    </h4>
-                    <div
-                        className={cx(styles.selectedGroup, {
-                            [styles.isCustom]: isCustom,
-                        })}
-                    >
-                        {isCustom ? 'Custom' : startCase(selectedGroupName)}
-                    </div>
-                </div>
-                <div
-                    // eslint-disable-next-line react/jsx-curly-brace-presence
-                    css={`
-                        align-items: center;
-                        display: flex;
-                    `}
-                >
-                    <Tooltip value="Remove">
-                        <Button
-                            kind="secondary"
-                            onClick={(event) => {
-                                event.stopPropagation()
-                                removeUser(userId)
-                            }}
-                            className={styles.button}
-                        >
-                            <SvgIcon name="trash" className={styles.trashIcon} />
-                        </Button>
-                    </Tooltip>
-                </div>
-            </div>
-            <animated.div className={styles.permissionControls} style={permissionControlsStyle}>
-                <div {...bind}>
-                    <RadioButtonGroup
-                        name={`UserPermissions${userId}`}
-                        className={cx(styles.groupSelector, {
-                            [styles.isSelected]: isSelected,
-                        })}
-                        options={permissionGroupOptions}
-                        onChange={(name) => {
-                            updatePermission(userId, State.getPermissionsForGroupName(resourceType, name))
-                        }}
-                        selectedOption={selectedGroupName}
-                        isCustom={isCustom}
-                    />
-                    <div className={styles.permissionsCheckboxes}>
-                        {Object.entries(userPermissions).map(([permission, value]) => (
-                            <div key={permission}>
-                                <div className={styles.checkboxContainer}>
-                                    <Checkbox
-                                        className={styles.checkbox}
-                                        id={`${userId}-${permission}`}
-                                        value={value}
-                                        onChange={() => updatePermission(userId, {
-                                            [permission]: !value,
-                                        })}
-                                    />
-                                </div>
-                                <label htmlFor={`${userId}-${permission}`}>
-                                    {startCase(I18n.t(`share.permissions.${permission}`))}
-                                </label>
-                            </div>
-                        ))}
-                    </div>
-                </div>
-            </animated.div>
-            {error && (
-                <div className={styles.errorMessage}>
-                    {error.message}
-                </div>
-            )}
-        </div>
-    )
-    /* eslint-enable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
-}
 
 function unsavedUnloadWarning(event) {
     const confirmationMessage = 'You have unsaved changes'
     const evt = (event || window.event)
     evt.returnValue = confirmationMessage // Gecko + IE
     return confirmationMessage // Webkit, Safari, Chrome etc.
-}
-
-function filterPermissions(permissions) {
-    // convert permission.anonymous to permission.user = 'anonymous', if needed
-    return permissions
-        .map((p) => State.fromAnonymousPermission(p))
-}
-
-/*
- * big horrible async handler for updating permission records
- * each permission that was added needs to send a request that it be created
- * each permission that was removed added needs to send a request that it be removed
- * Issues all requests in parallel, should not abort if one fails.
- * Returns any errors
- */
-
-async function savePermissions(currentUsers, props) {
-    const { resourceType, resourceId } = props
-    const oldPermissions = filterPermissions(await getResourcePermissions({
-        resourceType,
-        resourceId,
-    }))
-    const { added, removed } = State.diffUsersPermissions({
-        oldPermissions,
-        newUsers: currentUsers,
-        resourceType,
-    })
-
-    const allChangedUserIds = [
-        ...new Set([
-            ...added.map(({ user }) => user),
-            ...removed.map(({ user }) => user),
-        ]),
-    ]
-    const errors = {}
-    await Promise.all(allChangedUserIds.map(async (userId) => {
-        const userAddedItems = added.filter((p) => p.user === userId)
-        const userRemovedItems = removed.filter((p) => p.id != null && p.user === userId)
-        await Promise.all([
-            ...userAddedItems.map(async (data) => addResourcePermission({
-                resourceType,
-                resourceId,
-                data: State.toAnonymousPermission(data),
-            })),
-            ...userRemovedItems.map(async ({ id }) => removeResourcePermission({
-                resourceType,
-                resourceId,
-                id,
-            })),
-        ].map((task) => task.catch((error) => {
-            console.error(error) // eslint-disable-line no-console
-            // store failure but do not abort
-            // if user has multiple errors will only store one,
-            // this is fine, we don't need to show all of them
-            errors[userId] = error
-        })))
-    }))
-
-    return {
-        errors,
-    }
-}
-
-/**
- * Handling of permission state updates, selection & new user list
- * Factored out mainly to give main component some space.
- */
-
-function useUserPermissionState(props) {
-    const { resourceType, permissions: propPermissions } = props
-    const permissions = useMemo(() => filterPermissions(propPermissions), [propPermissions])
-    const users = useMemo(() => State.usersFromPermissions(resourceType, permissions), [resourceType, permissions])
-    const [currentUsers, setCurrentUsers] = useState(users) // state of user permissions to save to server
-    const [newUserIdList, setNewUserIdList] = useState([]) // users added since last save
-    const [selectedUserId, setSelectedUserId] = useState() // currently selected user
-
-    const canShareToUser = useCallback((userId) => (
-        // don't interfere with anonymous/current user
-        State.canShareToUser(userId)
-    ), [])
-
-    /* CRUD APIs */
-
-    const addUser = useCallback((userId) => {
-        if (!canShareToUser(userId)) { return }
-        setSelectedUserId(userId) // select new user on add
-        setCurrentUsers((prevUsers) => (
-            State.addUser(prevUsers, userId, State.getPermissionsForGroupName(resourceType, 'default'))
-        ))
-        // add/move user to start of new users. Remove before adding to start if already in array
-        setNewUserIdList((ids) => [userId, ...ids.filter((id) => id !== userId)])
-    }, [setCurrentUsers, canShareToUser, resourceType])
-
-    const removeUser = useCallback((userId) => {
-        if (!canShareToUser(userId)) { return }
-        setNewUserIdList((ids) => ids.filter((id) => id !== userId)) // remove user from new users list
-        setCurrentUsers((prevUsers) => State.removeUser(prevUsers, userId))
-    }, [setCurrentUsers, canShareToUser])
-
-    const updatePermission = useCallback((userId, permissions) => {
-        setCurrentUsers((prevUsers) => State.updatePermission(resourceType, prevUsers, userId, permissions))
-    }, [setCurrentUsers, resourceType])
-
-    // 'who has access' handler
-    const onAnonymousAccessChange = useCallback(({ value }) => {
-        const permissions = value === 'withLink'
-            ? State.getPermissionsForGroupName(resourceType, 'default')
-            : State.getEmptyPermissions(resourceType) // i.e. remove all permissions for anonymous
-        updatePermission('anonymous', permissions)
-    }, [updatePermission, resourceType])
-
-    // return everything
-
-    return useMemo(() => ({
-        permissions,
-        currentUsers,
-        canShareToUser,
-        addUser,
-        removeUser,
-        updatePermission,
-        onAnonymousAccessChange,
-        newUserIdList,
-        setNewUserIdList,
-        selectedUserId,
-        setSelectedUserId,
-    }), [
-        permissions,
-        currentUsers,
-        canShareToUser,
-        addUser,
-        removeUser,
-        updatePermission,
-        onAnonymousAccessChange,
-        newUserIdList,
-        setNewUserIdList,
-        selectedUserId,
-        setSelectedUserId,
-    ])
 }
 
 const ShareSidebar = connect(({ user }) => ({
@@ -761,25 +316,6 @@ const ShareSidebar = connect(({ user }) => ({
         /* eslint-enable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
     )
 })
-
-export function usePermissionsLoader({ resourceType, resourceId }) {
-    const loadPermissionsCallback = useCallback(() => (
-        getResourcePermissions({
-            resourceType,
-            resourceId,
-        })
-    ), [resourceType, resourceId])
-
-    const [loadState, loadPermissions] = useAsyncCallbackWithState(loadPermissionsCallback)
-    const hasProps = !!(resourceType && resourceId)
-    const { hasStarted } = loadState
-    useEffect(() => {
-        if (hasStarted || !hasProps) { return }
-        loadPermissions()
-    }, [hasStarted, hasProps, loadPermissions])
-
-    return [loadState, loadPermissions]
-}
 
 export default (props) => {
     const { resourceType, resourceId } = props
