+++ app/src/userpages/components/ShareSidebar/hooks/usePrevious.js
@@ -1,0 +1,9 @@
+import { useRef, useEffect } from 'react'
+
+export default function usePrevious(value) {
+    const ref = useRef()
+    useEffect(() => {
+        ref.current = value
+    }, [value])
+    return ref.current
+}
